tested signing up from console
