+++ main.py
@@ -1,0 +1,1 @@
+from application import app
